Refactor the command package to include Parser and CommandType
