+++ src/main/java/joni/Joni.java
@@ -1,6 +1,7 @@
 package joni;
 
 import joni.command.Command;
+import joni.command.Parser;
 import joni.task.TaskList;
 
 /**
